+++ packages/react/src/sds/ai/F0AiChat/components/MessagesContainer.tsx
@@ -167,7 +167,10 @@ const Messages = ({
               {turns.map((turnMessages, turnIndex) => (
                 <div
                   ref={turnIndex === turns.length - 1 ? lastTurnRef : undefined}
-                  className="flex flex-col items-start justify-start gap-2"
+                  className={cn(
+                    "flex flex-col items-start justify-start gap-2",
+                    turnIndex === turns.length - 1 && "pb-5"
+                  )}
                   key={`turn-${turnIndex}`}
                   style={{
                     minHeight:
@@ -260,6 +263,7 @@ const Messages = ({
             const callback =
               currentReaction === "like" ? onThumbsUp : onThumbsDown
             callback?.(message, { threadId, feedback })
+
             closeFeedbackModal()
           }}
           onClose={(message) => {